Split rendering methods into new files
